# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "description": "Documentation de l'API e-commerce.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/payments/initiate": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Initier un paiement",
                "description": "Initiate payment for an order",
                "operationId": "787ee735a11a5d5d3e904beefaf8962c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "payment_method",
                                    "amount"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "orange_money",
                                            "wave",
                                            "card"
                                        ]
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Order not eligible"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            }
        },
        "/payments/confirm": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Confirmer le paiement (simulation)",
                "description": "Confirm payment (webhook simulation)",
                "operationId": "f560558e9ccd2c97084f66665c39a49f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_id",
                                    "transaction_id",
                                    "status"
                                ],
                                "properties": {
                                    "payment_id": {
                                        "type": "string"
                                    },
                                    "transaction_id": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "success",
                                            "failed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Confirmé"
                    },
                    "400": {
                        "description": "Invalid payment ID"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            }
        },
        "/payments/{paymentId}/status": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Statut de paiement",
                "description": "Check payment status",
                "operationId": "61d8e20bad4a4e73cd0a3f88b22a95c1",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid payment ID"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            }
        },
        "/payments/simulate": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Simuler un paiement",
                "description": "Simulate payment callback (for testing)",
                "operationId": "53a3a80b252dcb23f8b1b21f43ca9729",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "success": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            }
        },
...
    },
...
}